+++ config/prod.exs
@@ -1,7 +1,6 @@
 use Mix.Config
 
 config :dealog_backoffice, DealogBackofficeWeb.Endpoint,
-  url: [host: {:system, "HOSTNAME"}, port: 443, scheme: "https"],
   cache_static_manifest: "priv/static/cache_manifest.json",
   server: true,
   version: Application.spec(:dealog_backoffice, :vsn)