+++ config/releases.exs
@@ -45,14 +45,20 @@ secret_key_base =
     You can generate one by calling: mix phx.gen.secret
     """
 
+hostname =
+  System.get_env("HOSTNAME") ||
+    raise """
+    environment variable HOSTNAME is missing.
+    """
+
 config :dealog_backoffice, DealogBackofficeWeb.Endpoint,
+  url: [scheme: "https", host: hostname, port: 443],
   http: [
     port: String.to_integer(System.get_env("PORT") || "4000"),
     transport_options: [socket_opts: [:inet6]]
   ],
+  check_origin: ["//#{hostname}"],
   secret_key_base: secret_key_base
-
-config :dealog_backoffice, DealogBackofficeWeb.Endpoint, server: true
 
 smtp_relay =
   System.get_env("SMTP_RELAY") ||